Add files after exam on 23 july

+++ src/com/oca_ocp/oca_ch6_working_methods_encapsulation/FootBall.java
@@ -2,14 +2,14 @@ package com.oca_ocp.oca_ch6_working_methods_encapsulation;
 
 public class FootBall {
 
-    public static Long getScore(long TimeRemaining) { // Long will not compile
+    public static Long getScore(Long TimeRemaining) { // Long will not compile
 
         return 2*TimeRemaining;
     }
 
     public static void main(String[] args) {
 
-        final int startTime = 2;
+        final long startTime = 2;
         getScore(startTime);
     }
 }
